submissions/submission_multilingual_hateclipper/requirements.txt
torch
torchvision
pytesseract
opencv-python-headless
scipy
scikit-image
ninja
numpy
easyocr==1.6.2
tqdm
pyyaml
transformers
pytorch_lightning
pillow
multilingual_clip